Cleanup test organization (#296)
Closes #183.

Creates a cohesive structure for tests by splitting class tests into folders, all PyGraph class
tests are in the tests/graph folder and all PyDiGraph/PyDAG tests are in the tests/DiGraph
folder. As before, all retworkx.generators tests are in the tests/generators folder.

There are two exceptions:

 * test_dispatch.py is in the tests folder because it tests both PyGraph and PyDigraph and
   moving this test to separate folders would lead to duplicated code.
 * test_random.py is in the tests folder because it is similar to generators and generates
   random graphs, however they are not in the generators module, so I was unsure if it is
   needed to create a random folder to put this kind of tests.

* Moved test_adj to folder

Split folders for PyGraph an PyDigraph/PyDAG

* Move adjencency matrix tests for Pygraph

* Move adjencency matrix for DiGraph

* Move all simple paths for Pygraph

* Move all simple paths for PyDigraph

* Move test ancestors and descendants to PyDigraph

* Move test astar to PyDigraph

* Move Astar to Graph

* Move test collect runs to DiGraph

* Move test compose to DiGraph

* Move compose to graph

* Move test custom return types to DiGraph

* Move test cycle basis to Graph

* Move test deepcopy to DiGraph

* Move test depth to DiGraph

* Move test dfs to DiGraph

* Move dfs to Graph

* Move test dijkstra to Digraph

* Move test dijkstra to graph

* Move test dispatch to DiGraph

* Move test dispatch to Graph

* Move test distance matrix to Digraph

* Move test distance matrix to Graph

* Move test dot to Digraph

* Move test dot to Graph

* Move test edgelist to Digraph

* Move test edgelist to Graph

* Move test edges to DiGraph

* Move test find cycle to DiGraph

* Move test floyd warshall to DiGraph

* Move test floyd warshall to Graph

* Move test isomorphic to PyDigraph

* Move test k shortest path to DiGraph

* Move test k shortest path to Graph

* Move test layers to DiGraph

* Move test matching to Graph

* move max weight matching to Graph

* move test_unique_neighbors_on_graphs to Graph

* Move test neighbors to DiGraph

* Move test nodes to DiGraph

* move test predecessor and sucessor to DiGraph

* move test strongly connected to DiGraph

* move test subgraph to pygraph

* move test symmetric to DiGraph

* move test to undirected to DiGraph

* move test union to DiGraph

* move test weakly connected to DiGraph

* remove test dispatch

Removed test dispatch because it is better to keep it in the top level in order to not repeat unecessary code

* lint

* change digraph folder name

* Move new test files into new structure
