+++ tests/digraph/test_adj.py
