+++ tests/digraph/test_core_number.py
@@ -66,37 +66,6 @@ class TestCoreNumber(unittest.TestCase):
         example_core[20] = 0
         self.example_core = example_core
 
-    def test_undirected_empty(self):
-        graph = retworkx.PyGraph()
-        res = retworkx.core_number(graph)
-        self.assertIsInstance(res, dict)
-        self.assertEqual(res, {})
-
-    def test_undirected_all_0(self):
-        graph = retworkx.PyGraph()
-        graph.add_nodes_from(list(range(4)))
-        res = retworkx.core_number(graph)
-        self.assertIsInstance(res, dict)
-        self.assertEqual(res, {0: 0, 1: 0, 2: 0, 3: 0})
-
-    def test_undirected_all_3(self):
-        graph = retworkx.PyGraph()
-        graph.add_nodes_from(list(range(4)))
-        graph.add_edges_from_no_data([
-            (0, 1), (0, 2), (0, 3), (1, 2), (1, 3), (2, 3)
-        ])
-        res = retworkx.core_number(graph)
-        self.assertIsInstance(res, dict)
-        self.assertEqual(res, {0: 3, 1: 3, 2: 3, 3: 3})
-
-    def test_undirected_paper_example(self):
-        graph = retworkx.PyGraph()
-        graph.add_nodes_from(list(range(21)))
-        graph.add_edges_from_no_data(self.example_edges)
-        res = retworkx.core_number(graph)
-        self.assertIsInstance(res, dict)
-        self.assertEqual(res, self.example_core)
-
     def test_directed_empty(self):
         digraph = retworkx.PyDiGraph()
         res = retworkx.core_number(digraph)
@@ -104,7 +73,7 @@ class TestCoreNumber(unittest.TestCase):
         self.assertEqual(res, {})
 
     def test_directed_all_0(self):
-        digraph = retworkx.PyGraph()
+        digraph = retworkx.PyDiGraph()
         digraph.add_nodes_from(list(range(4)))
         res = retworkx.core_number(digraph)
         self.assertIsInstance(res, dict)
