+++ tests/digraph/test_custom_return_types.py
