+++ tests/digraph/test_depth.py
