+++ tests/digraph/test_nodes.py
