+++ tests/digraph/test_pred_succ.py
