+++ tests/digraph/test_strongly_connected.py
