+++ tests/digraph/test_symmetric.py
